ocr_pdf2ppt_v3/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-pdf2ppt-v3"
version = "0.3.0"
description = "Isolated V3 PDF to editable PPT converter with new renderer/compositor/quality gate"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Codex" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.5.0",
  "openai>=1.50.0",
  "pymupdf>=1.24.0",
  "python-pptx>=1.0.2",
  "Pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
paddle = [
  "paddleocr[doc-parser]>=3.4.0",
  "paddlepaddle>=3.3.0",
]

[project.scripts]
ocr-pdf2ppt-v3 = "ocr_pdf2ppt_v3.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
